// generators/src/types.ts
export type EspansoType = "date" | "choice" | "clipboard" | "echo" | "shell" | "random";

export type EspansoShell = "Powershell" | "bash" | "sh"

export type EspansoTrigger = `:${string}` | Array<`:${string}`>;

export type EspansoVarReplacer = {
    trigger: EspansoTrigger;
    replace: `{{${string}}}` | string;
    vars?: [
        {
            name: string;
            type: EspansoType;
            params: Partial<{
                shell: EspansoShell;
                cmd: string | string[];
                format: string;
                choices: string[];
                locale: string;
                echo: string;
                values: Array<string | { label: string; id: string }>;
            }>;
        },
    ];
};